git-svn-id: svn+ssh://rubyforge.org/var/svn/myexperiment/trunk@1802 4f7b4649-9ca9-420a-81b3-55a188e4d46f

+++ app/models/workflow.rb
@@ -14,18 +14,19 @@ require 'acts_as_runnable'
 
 require 'scufl/model'
 require 'scufl/parser'
-require 'scufl/dot'
 
 class Workflow < ActiveRecord::Base
+  
   has_many :citations, 
            :order => "created_at DESC",
            :dependent => :destroy
 
-  belongs_to :content_blob 
-
-  # need to destroy the content_blobs belonging to the workflow versions to avoid orphaned records
+  belongs_to :content_blob
+  
+  # need to destroy the workflow versions and their content blobs to avoid orphaned records
   before_destroy { |w| w.versions.each do |wv|
-                        wv.content_blob.destroy
+                        wv.content_blob.destroy if wv.content_blob
+                        wv.destroy
                       end }
 
   acts_as_contributable
@@ -37,11 +38,13 @@ class Workflow < ActiveRecord::Base
   
   acts_as_reviewable
 
-  explicit_versioning(:version_column => "current_version", :file_columns => ["image", "svg"], :white_list_columns => ["body"]#, :other_columns => ["content_blob_id"]
-  ) do
+  explicit_versioning(:version_column => "current_version", 
+                      :file_columns => ["image", "svg"], 
+                      :white_list_columns => ["body"]) do
+    
     file_column :image, :magick => {
       :versions => {
-        :thumb    => { :size => "100x100!" }, 
+        :thumb    => { :size => "100x100" }, 
         :medium   => { :size => "500x500>" },
         :full     => { }
       }
@@ -52,9 +55,14 @@ class Workflow < ActiveRecord::Base
     format_attribute :body
     
     belongs_to :content_blob
+    
     # :dependent => :destroy is not supported in belongs_to in rails 1.2.6
-    after_destroy { |wv| wv.content_blob.destroy }
-     
+    after_destroy { |wv| wv.content_blob.destroy if wv.content_blob }
+    
+    # Update the parent contribution model buy only if this isn't the current version (because the workflow model will take care of that).
+    # This is required to keep the contribution's updated_at field accurate.
+    after_save { |wv| wv.workflow.contribution.save if wv.workflow.contribution && wv.version != wv.workflow.current_version }
+    
   end
   
   #non_versioned_fields.push("image", "svg", "license", "tag_list") # acts_as_versioned and file_column don't get on
@@ -71,18 +79,18 @@ class Workflow < ActiveRecord::Base
   
   format_attribute :body
   
+  validates_presence_of :unique_name
   validates_uniqueness_of :unique_name
   
   validates_inclusion_of :license, :in => [ "by-nd", "by-sa", "by" ]
+  
+  validates_presence_of :content_type
 
   file_column :image, :magick => {
     :versions => {
-      :thumb    => { :size => "100x100!" }, 
+      :thumb    => { :size => "100x100>" }, 
       :medium   => { :size => "500x500>" },
       :full     => { },
-      :padlock => { :transformation => Proc.new { |image| image.resize(100, 100).blur_image.composite(Magick::ImageList.new("#{RAILS_ROOT}/public/images/padlock.gif"), 
-                                                                                                      Magick::SouthEastGravity, 
-                                                                                                      Magick::OverCompositeOp) } }
     }
   }
   
@@ -100,56 +108,61 @@ class Workflow < ActiveRecord::Base
     return list
   end
   
-  # Begin SCUFL specific methods
-
-  def get_input_ports(version)
-    return nil unless (workflow_version = self.find_version(version))
-    parser = Scufl::Parser.new
-    model  = parser.parse(workflow_version.content_blob.data)
-    
-    return model.sources
-  end
-  
-  def get_sculf_model(version)
-    return nil unless (workflow_version = self.find_version(version))
-    parser = Scufl::Parser.new
-    model  = parser.parse(workflow_version.content_blob.data)
-    
-    return model
-  end
-
-  # End SCUFL specific methods
-
-  def named_download_url
-    "#{BASE_URI}/workflows/#{id}/download/#{unique_name}.xml"
-  end
-
-  def create_workflow_diagrams(scufl_model, extension)
-
-    salt = rand 32768
-
-    self.title = scufl_model.description.title.blank? ? "untitled" : scufl_model.description.title
-    self.unique_name = "#{self.title.gsub(/[^\w\.\-]/,'_').downcase}_#{salt}"
-
-    unless RUBY_PLATFORM =~ /mswin32/
-      i = Tempfile.new("image")
-      Scufl::Dot.new.write_dot(i, scufl_model)
-      i.close(false)
-      img = StringIO.new(`dot -Tpng #{i.path}`)
-      svg = StringIO.new(`dot -Tsvg #{i.path}`)
-      img.extend FileUpload
-      img.original_filename = "#{self.unique_name}_#{extension}.png"
-      img.content_type = "image/png"
-      svg.extend FileUpload
-      svg.original_filename = "#{self.unique_name}_#{extension}.svg"
-      svg.content_type = "image/svg+xml"
-      self.image = img
-      self.svg = svg
+  # Sets an internal unique name for this workflow.
+  def set_unique_name
+    salt = rand 1000000
+    salt2 = rand 100
+    if self.title.blank?
+      self.unique_name = "#{salt}_#{salt2}"        
+    else
+      self.unique_name = "#{self.title.gsub(/[^\w\.\-]/,'_').downcase}_#{salt}"
     end
   end
+  
+  def processor_class
+    @processor_class ||= WorkflowTypesHandler.processor_class_for_content_type(self.content_type)
+  end
+  
+  def can_infer_metadata_for_this_type?
+    proc_class = self.processor_class
+    return false if proc_class.nil?
+    return proc_class.can_infer_metadata?
+  end
+  
+  def type_display_name
+    WorkflowTypesHandler.type_display_name_for_content_type(self.content_type)  
+  end
+  
+  def display_data_format
+    klass = self.processor_class
+    @display_data_format = (klass.nil? ? self.file_ext : klass.display_data_format)
+  end
+  
+  def get_workflow_model_object(version)
+    return nil unless (workflow_version = self.find_version(version))
+    return (self.processor_class.nil? ? nil : self.processor_class.new(workflow_version.content_blob.data).get_workflow_model_object)
+  end
+  
+  # Begin acts_as_runnable overridden methods
+ 
+  def get_input_ports(version)
+    return nil unless (workflow_version = self.find_version(version))
+    return (self.processor_class.nil? ? nil : self.processor_class.new(workflow_version.content_blob.data).get_workflow_model_input_ports)
+  end
+  
+  # End acts_as_runnable overridden methods
 
-end
+  def filename(version=nil)
+    if version.blank?
+      return "#{unique_name}.#{file_ext}"
+    else
+      return nil unless (workflow_version = self.find_version(version))
+      return "#{workflow_version.unique_name}.#{file_ext}"
+    end
+  end
+  
+  def named_download_url
+    "#{BASE_URI}/workflows/#{id}/download/#{filename}"
+  end
 
-module FileUpload
-  attr_accessor :original_filename, :content_type
 end
